+++ app/Api/Controllers/UsersController.php
@@ -5,7 +5,7 @@ namespace App\Api\Controllers;
 use App\Api\Core\Traits\ResponseTrait;
 use App\Api\Repositories\UsersRepository;
 use App\Api\Requests\CreateUserRequest;
-use App\Api\Resources\UsersResource;
+use App\Api\Resources\UserResource;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
@@ -24,13 +24,13 @@ class UsersController extends BaseController
     public function createCustomer(CreateUserRequest $request)
     {
         try {
-            return $this->success('Created customer', new UsersResource($this->usersRepository->createCustomer($request)));
+            return $this->success('Created customer', new UserResource($this->usersRepository->createCustomer($request)));
         } catch (\Exception $e) {
-           return $this->error($e->getMessage(), $e->getCode());
+            return $this->error($e->getMessage(), $e->getCode());
         }
     }
 
     public function get(Request $request){
-        return $this->success('Returned user', new UsersResource($this->usersRepository->getUserById($request->user()->id)));
+        return $this->success('Returned user', new UserResource($request->user()));
     }
 }